feat(scripts): detect test binary panics in bazel output analyzer
- Fix section delimiter regex to only match start markers
- Fix section pairing to use consecutive markers with virtual EOF
- Detect panic: in target output sections (.panic suffix)
- Add parse_bazel_go_test_panic_cases to report crash info
  to bazel-go-test-problem-cases.json (shard, cases, panic, flaky_attempts)
- Add panic: to indexReg for visibility in bazel-go-test-index.log
- Address review: jq injection via --arg, O(N^2) to O(N), break->continue,
  validate log file before wc -l, check jq availability,
  configurable MAX_PANIC_LENGTH, deprecation notices
- Deprecate report-flaky-cases.ts and report_job_result.sh
